Split worktree detail toolbar views
